revert: "feat: move `ProjectView` to `ApiWrapper`"
This reverts commit c6d3c0bc37bcccab0a3ac866c748a7bd2a447acb.
