--- ModrinthApiWrapper/src/main/java/me/theclashfruit/mrapi/serializers/ModrinthDonationUrlsModel.kt
@@ -1,10 +1,0 @@
-package me.theclashfruit.mrapi.serializers
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ModrinthDonationUrlsModel (
-  var id       : String? = null,
-  var platform : String? = null,
-  var url      : String? = null
-)
