--- ModrinthApiWrapper/src/main/java/me/theclashfruit/mrapi/serializers/ModrinthGalleryModel.kt
@@ -1,12 +1,0 @@
-package me.theclashfruit.mrapi.serializers
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ModrinthGalleryModel (
-  var url         : String?  = null,
-  var featured    : Boolean? = null,
-  var title       : String?  = null,
-  var description : String?  = null,
-  var created     : String?  = null
-)
