--- ModrinthApiWrapper/src/main/java/me/theclashfruit/mrapi/serializers/ModrinthLicenseModel.kt
@@ -1,10 +1,0 @@
-package me.theclashfruit.mrapi.serializers
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ModrinthLicenseModel (
-  var id   : String? = null,
-  var name : String? = null,
-  var url  : String? = null
-)
