--- ModrinthApiWrapper/src/main/java/me/theclashfruit/mrapi/serializers/ModrinthModMessageModel.kt
@@ -1,9 +1,0 @@
-package me.theclashfruit.mrapi.serializers
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ModrinthModMessageModel(
-    var message : String? = null,
-    var body    : String? = null
-)
